python/cloudfront-s3/requirements.txt
aws-cdk.aws-cloudfront-origins
aws-cdk.aws-s3-deployment
aws-cdk.aws-s3
aws-cdk.aws-s3-assets